new object code structure org
