+++ include/Compiler.h
@@ -2,4 +2,6 @@
 #define WIND_COMPILER_H
 // Header that is responsible for compiling source code into instructions.
 
+#include "WindObject.h"
+
 #endif
